Update BreastTomography.py
I have fixed two issues in the code provided:

The computation of the step size at each iteration of the conjugate gradient algorithm initially assumed that the grids over which (1) the plane waves were propagated and (2) the sound speed was reconstructed were the same. However, we interpolate the backprojection in (1) onto the grid in (2). To account for the change in the voxel size we added a grid_conv_factor to both the MATLAB and Python code.
The code initially provided for the work presented at SPIE Medical Imaging 2022 assumed that the source waveform was known exactly in both shape and amplitude. However, it is practically impossible to know the source waveform exactly, so source estimation must be automated. I've provided an implementation of source estimation (requested by zero34521 above) as a frequency-dependent complex-valued scaling factor that can vary with each transmission. The updated code tests the waveform inversion algorithm with source estimation by applying a random complex-valued scaling to each frequency in the transmitted waveform. However, automated source estimation also makes waveform inversion less robust to cycle skipping, so we used lower frequencies in our waveform. I will discuss this tradeoff between automated source estimation and cycle skipping in a future work.

+++ Python/BreastTomography.py
@@ -17,7 +17,7 @@ array_separation = data_in['array_separation'][0,0];
 del data_in;
 
 # Time Axis and Sampling Frequency
-freq_bins_used = np.arange(0,(f.size+1)/2,10); # Frequency Bins Used
+freq_bins_used = np.arange(0,int((f.size+1)/2),10); # Frequency Bins Used
 fused = f[freq_bins_used]; # Frequencies Used
 P_fused = P_f[freq_bins_used]; # Frequency Bins in Pulse
 dt = np.mean(np.diff(t)); fs = 1/dt; # Sampling Period [s] and Frequency [Hz]
